+++ CHANGELOG.md
@@ -5,7 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.9.4] - 2026-02-06
+
+### Changed
+
+- `Dealloc` traits' fallible functions now treat zero-sized layouts and dangling pointers as a hard
+  error
+- `Dealloc` traits' fallible functions are now a noop for ZSLs and dangling pointers
+- All other allocation functions now treat ZSLs as an error.
 
 ## [0.9.2] - 2026-02-03